medical-image-segmentation/requirements.txt
streamlit==1.32.0
numpy==1.26.4
opencv-python==4.9.0.80
Pillow==10.1.0
matplotlib==3.8.2
scikit-learn==1.4.0
scikit-image==0.22.0
torch==2.2.0
torchvision==0.17.0
pydensecrf==1.0rc3
